instagram-to-telegram/requirements.txt
anyio==4.0.0
certifi==2023.7.22
charset-normalizer==3.3.2
exceptiongroup==1.1.3
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
idna==3.4
instagrapi==2.0.0
Pillow==10.1.0
psycopg2-binary==2.9.9
pycryptodomex==3.18.0
pydantic==1.10.9
PySocks==1.7.1
python-decouple==3.8
python-telegram-bot==20.6
requests==2.31.0
sniffio==1.3.0
typing_extensions==4.8.0
urllib3==2.0.7
speedtest-cli==2.1.3
